+++ src/context/index.js
@@ -1,0 +1,18 @@
+import favoriteContext, { FavoriteProvider } from './favorite'
+import historyContext, { HistoryProvider } from './history'
+import homeContext, { HomeProvider } from './home'
+import resultsContext, { ResultsProvider } from './results'
+import searchContext, { SearchProvider } from './search'
+
+export {
+  favoriteContext,
+  FavoriteProvider,
+  historyContext,
+  HistoryProvider,
+  homeContext,
+  HomeProvider,
+  resultsContext,
+  ResultsProvider,
+  searchContext,
+  SearchProvider,
+}
